+++ src/gui/native_controls/macros.rs
@@ -79,7 +79,7 @@ macro_rules! ctrlid_hwnd_on_onsubclass {
 		hwnd_on_onsubclass!($evstruc);
 
 		/// Returns the control ID.
-		pub fn ctrl_id(&self) -> u16 {
+		pub fn ctrl_id(&self) -> i32 {
 			match &self.0.opts_id {
 				OptsId::Wnd(opts) => opts.ctrl_id,
 				OptsId::Dlg(ctrl_id) => *ctrl_id,
